Added display for game data

+++ utils.py
@@ -2,8 +2,8 @@ import settings
 
 
 def height_prct(percentage):
-    return settings.HEIGHT * (percentage / 100)
+    return int(settings.HEIGHT * (percentage / 100))
 
 
 def width_prct(percentage):
-    return settings.WIDTH * (percentage / 100)
+    return int(settings.WIDTH * (percentage / 100))
